requirements.txt
Flask
pandas
numpy<2.0.0
tensorflow-cpu==2.20.0
joblib
scikit-learn
gunicorn